Leaf Classifier
This project focuses on automated classification of various leaf types using a Convolutional Neural Network (CNN) implemented in TensorFlow and Keras. The goal is to accurately identify plant species by analyzing the unique visual patterns of their leaves.
To achieve this, a dataset of leaf images was preprocessed through resizing, normalization, and augmentation to improve model robustness and generalization. The CNN model architecture includes multiple convolutional and pooling layers followed by dense layers to effectively capture spatial hierarchies in the leaf structures.To ensure smooth deployment and cross-platform consistency, the entire application was Dockerized. This allows the trained model to be packaged with all its dependencies, ensuring that it runs identically across different systems and environments.